+++ packages/datagrid/src/constant.js
@@ -1,3 +1,0 @@
-const I18N_DOMAIN_DATAGRID = 'tui-datagrid';
-
-export default I18N_DOMAIN_DATAGRID;