+++ packages/datagrid/src/index.ts
@@ -1,0 +1,7 @@
+import * as constants from './constants';
+import DataGrid, * as components from './components';
+
+export default Object.assign(DataGrid, {
+	constants,
+	components,
+});