+++ packages/datagrid/stories/DynamicDataGrid.component.js
@@ -3,10 +3,9 @@ import random from 'lodash/random';
 import { IconsProvider } from '@talend/react-components';
 import PropTypes from 'prop-types';
 
-import DataGrid from '../src/components/';
+import DataGrid from '../src/components';
 import serializer from '../src/components/DatasetSerializer';
 import sample from './sample.json';
-import getComponent from './getComponent';
 
 const ADD_ITEMS_NUMBER = 4;
 const LOADING_ITEM = {
@@ -40,24 +39,6 @@ export function getRowDataInfos(rowData) {
 			loading: 0,
 			notLoaded: 0,
 		},
-	);
-}
-
-/**
- * redrawRows - call redrawRows only if necessary. (improve ag-grid performance)
- *
- * @param  {object} props     component props
- * @param  {object} prevProps previous component prop
- * @return {boolean}         return true if we need to reload the grid
- */
-function forceRedrawRows(props, prevProps) {
-	const prevInfos = getRowDataInfos(prevProps.rowData);
-	const currentInfos = getRowDataInfos(props.rowData);
-
-	return (
-		prevInfos.loaded !== currentInfos.loaded ||
-		prevInfos.loading !== currentInfos.loading ||
-		prevInfos.notLoaded !== currentInfos.notLoaded
 	);
 }
 
@@ -111,9 +92,7 @@ function getItemWithRandomValue() {
 }
 
 export default class DynamicDataGrid extends React.Component {
-	static propTypes = {
-		forceRedraw: PropTypes.bool,
-	};
+	static propTypes = {};
 
 	constructor() {
 		super();
@@ -122,7 +101,7 @@ export default class DynamicDataGrid extends React.Component {
 		this.terminateItems = this.terminateItems.bind(this);
 		this.changeColumnQuality = this.changeColumnQuality.bind(this);
 
-		const datagridSample = Object.assign({}, sample);
+		const datagridSample = { ...sample };
 		datagridSample.data = new Array(ADD_ITEMS_NUMBER).fill().map(() => ({ ...LOADING_ITEM }));
 		this.state = { sample: datagridSample, loading: true, index: 1 };
 
@@ -131,7 +110,7 @@ export default class DynamicDataGrid extends React.Component {
 
 	terminateItems() {
 		this.setState(prevState => {
-			const datagridSample = Object.assign({}, prevState.sample);
+			const datagridSample = { ...prevState.sample };
 			datagridSample.data.splice(
 				(prevState.index - 1) * ADD_ITEMS_NUMBER,
 				ADD_ITEMS_NUMBER,
@@ -147,7 +126,7 @@ export default class DynamicDataGrid extends React.Component {
 
 	changeColumnQuality() {
 		this.setState(prevState => {
-			const datagridSample = Object.assign({}, prevState.sample);
+			const datagridSample = { ...prevState.sample };
 			datagridSample.data = datagridSample.data.map(rowData => ({
 				...rowData,
 				value: {
@@ -166,7 +145,7 @@ export default class DynamicDataGrid extends React.Component {
 	}
 
 	addLoadingsItems() {
-		const datagridSample = Object.assign({}, this.state.sample);
+		const datagridSample = { ...this.state.sample };
 		datagridSample.data = datagridSample.data.concat(
 			new Array(ADD_ITEMS_NUMBER).fill().map(() => ({ ...LOADING_ITEM })),
 		);
@@ -191,19 +170,13 @@ export default class DynamicDataGrid extends React.Component {
 				/>
 				<input
 					type="button"
-					value={'change quality for 1 column(#6)'}
+					value="change quality for 1 column(#6)"
 					onClick={this.changeColumnQuality}
 					disabled={this.state.loading}
 				/>
 				Number of data : {this.state.sample.data.length}
 				<IconsProvider />
-				<DataGrid
-					data={this.state.sample}
-					getComponent={getComponent}
-					rowSelection="multiple"
-					forceRedrawRows={this.props.forceRedraw ? forceRedrawRows : null}
-					rowData={this.props.forceRedraw ? serializer.getRowData(this.state.sample) : null} //
-				/>
+				<DataGrid data={this.state.sample} rowSelection="multiple" rowData={null} />
 			</div>
 		);
 	}