+++ tools/scripts-config-jest/test-setup.js
@@ -2,10 +2,23 @@ require('@testing-library/jest-dom');
 require('core-js/stable');
 require('regenerator-runtime/runtime');
 require('raf/polyfill');
-const configure = require('enzyme').configure;
-const Adapter = require('enzyme-adapter-react-16');
 
-configure({ adapter: new Adapter() });
+// enzyme adapter configuration
+let React;
+try {
+	React = require('react');
+} catch (e) {}
+
+const version = React && React.version;
+if (version && version.startsWith('16.')) {
+	const configure = require('enzyme').configure;
+	const Adapter = require('enzyme-adapter-react-16');
+	configure({ adapter: new Adapter() });
+} else if (version && version.startsWith('17.')) {
+	const configure = require('enzyme').configure;
+	const Adapter = require('@wojtekmaj/enzyme-adapter-react-17');
+	configure({ adapter: new Adapter() });
+}
 
 // Mock fetch
 const fetch = jest.fn(