+++ packages/components/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @talend/react-components
 
+## 6.46.6
+
+### Patch Changes
+
+- 0332ed371: Datalist : Blur should persist value only if the filter value is different from the selected one
+
+## 6.46.5
+
+### Patch Changes
+
+- c834a7ec6: fix(Notification): remove this in functional component
+- Updated dependencies [453802c72]
+- Updated dependencies [6f68d11f4]
+  - @talend/design-system@2.6.0
+  - @talend/icons@6.41.0
+
+## 6.46.4
+
+### Patch Changes
+
+- 344a7551f: fix: proptypes of VList
+
+## 6.46.3
+
+### Patch Changes
+
+- 3be3e132d: Datalist should update its value when the titleMap changes
+
+## 6.46.2
+
+### Patch Changes
+
+- e04e3910f: feat(components/QualityBar): rework quality bar with design tokens
+
+## 6.46.1
+
+### Patch Changes
+
+- bcc5e70ed: fix(VList): handle VList without columns
+
 ## 6.46.0
 
 ### Minor Changes